[react-events] Separate the Focus/FocusWithin unit tests (#16298)
